requirements.txt
jsonschema==2.6.0
singer-python==5.6.1
boto3==1.18.15
